+++ espnet_onnx/export/asr/models/__init__.py
@@ -1,29 +1,18 @@
 # decoder
 from espnet2.asr.decoder.rnn_decoder import RNNDecoder as espnetRNNDecoder
-# from espnet2.asr.encoder.conformer_encoder import \
-#     ConformerEncoder as espnetConformerEncoder
 from espnet2.asr.encoder.contextual_block_conformer_encoder import \
     ContextualBlockConformerEncoder as espnetContextualConformer
 from espnet2.asr.encoder.contextual_block_transformer_encoder import \
     ContextualBlockTransformerEncoder as espnetContextualTransformer
-# encoder
-# from espnet2.asr.encoder.rnn_encoder import RNNEncoder as espnetRNNEncoder
-# from espnet2.asr.encoder.transformer_encoder import \
-#     TransformerEncoder as espnetTransformerEncoder
-# from espnet2.asr.encoder.vgg_rnn_encoder import \
-#     VGGRNNEncoder as espnetVGGRNNEncoder
 
-from espnet_onnx.export.asr.models.ctc import CTC
+
+from espnet_onnx.export.asr.models.layers.ctc import CTC
 from espnet_onnx.export.asr.models.decoders.rnn import RNNDecoder
 from espnet_onnx.export.asr.models.decoders.transducer import TransducerDecoder
 from espnet_onnx.export.asr.models.decoders.xformer import XformerDecoder
-# from espnet_onnx.export.asr.models.encoders.conformer import ConformerEncoder
 from espnet_onnx.export.asr.models.encoders.contextual_block_xformer import \
     ContextualBlockXformerEncoder
-# from espnet_onnx.export.asr.models.encoders.rnn import RNNEncoder
-# from espnet_onnx.export.asr.models.encoders.transformer import \
-#     TransformerEncoder
-from espnet_onnx.export.asr.models.joint_network import JointNetwork
+from espnet_onnx.export.asr.models.layers.joint_network import JointNetwork
 
 try:
     from espnet2.asr.transducer.transducer_decoder import \
@@ -69,10 +58,6 @@ def get_encoder(model, frontend, preencoder, export_config, convert_map):
             export_config=export_config
         )
         return DefaultEncoder(_model, frontend, **export_config)
-    # elif isinstance(model, espnetTransformerEncoder):
-    #     return TransformerEncoder(model, frontend, preencoder, **export_config)
-    # elif isinstance(model, espnetConformerEncoder):
-    #     return ConformerEncoder(model, frontend, preencoder, **export_config)
 
 
 def get_decoder(model, export_config):
