+++ espnet_onnx/export/asr/models/decoders/predecoder.py
@@ -4,7 +4,7 @@ import torch
 import torch.nn as nn
 from espnet.nets.pytorch_backend.rnn.attentions import NoAtt
 
-from espnet_onnx.export.layers.attention import require_tanh
+from espnet_onnx.export.asr.models.layers.attention import require_tanh
 from espnet_onnx.utils.abs_model import AbsExportModel
