+++ espnet_onnx/export/asr/models/encoders/conformer.py
@@ -1,6 +1,4 @@
-import os
 
-import torch
 import torch.nn as nn
 from espnet.nets.pytorch_backend.transformer.attention import (
     LegacyRelPositionMultiHeadedAttention, MultiHeadedAttention,
@@ -9,11 +7,9 @@ from espnet.nets.pytorch_backend.transformer.subsampling import (
     Conv2dSubsampling, Conv2dSubsampling2, Conv2dSubsampling6,
     Conv2dSubsampling8)
 
-from espnet_onnx.export.asr.get_config import (get_frontend_config,
-                                               get_norm_config)
-from espnet_onnx.export.asr.models.conformer_layer import OnnxConformerLayer
-from espnet_onnx.export.asr.models.language_models.embed import Embedding
-from espnet_onnx.export.asr.models.multihead_att import (
+from espnet_onnx.export.asr.models.layers.conformer_layer import OnnxConformerLayer
+from espnet_onnx.export.asr.models.layers.embed import Embedding
+from espnet_onnx.export.asr.models.layers.multihead_att import (
     OnnxMultiHeadedAttention, OnnxRelPosMultiHeadedAttention)
 from espnet_onnx.utils.abs_model import AbsExportModel
 from espnet_onnx.utils.torch_function import MakePadMask
