+++ espnet_onnx/export/asr/models/encoders/rnn.py
@@ -1,17 +1,12 @@
-import os
 
 import six
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from espnet2.asr.encoder.vgg_rnn_encoder import \
-    VGGRNNEncoder as espnetVGGRNNEncoder
 from espnet.nets.pytorch_backend.rnn.encoders import RNN, RNNP, VGG2L
-
-from espnet_onnx.export.asr.get_config import (get_frontend_config,
-                                               get_norm_config)
-from espnet_onnx.utils.abs_model import AbsExportModel
 from espnet.nets.pytorch_backend.nets_utils import make_pad_mask
+
+from espnet_onnx.utils.abs_model import AbsExportModel
 
 
 class OnnxRNNP(nn.Module):
