+++ espnet_onnx/export/asr/models/language_models/transformer.py
@@ -8,9 +8,9 @@ from espnet.nets.pytorch_backend.transformer.attention import \
 from espnet.nets.pytorch_backend.transformer.subsampling import (
     Conv2dSubsampling, Conv2dSubsampling6, Conv2dSubsampling8)
 
-from espnet_onnx.export.asr.models.encoder_layer import OnnxEncoderLayer
-from espnet_onnx.export.asr.models.language_models.embed import Embedding
-from espnet_onnx.export.asr.models.multihead_att import \
+from espnet_onnx.export.asr.models.layers.encoder_layer import OnnxEncoderLayer
+from espnet_onnx.export.asr.models.layers.embed import Embedding
+from espnet_onnx.export.asr.models.layers.multihead_att import \
     OnnxMultiHeadedAttention
 from espnet_onnx.utils.abs_model import AbsExportModel
 from espnet_onnx.utils.torch_function import MakePadMask
