+++ espnet_onnx/export/tts/models/tts_models/tacotron2.py
@@ -8,8 +8,8 @@ import torch.nn as nn
 import torch.nn.functional as F
 from espnet.nets.pytorch_backend.rnn.attentions import AttForward, AttForwardTA
 
-from espnet_onnx.export.layers.attention import get_attention
-from espnet_onnx.export.layers.predecoder import PreDecoder
+from espnet_onnx.export.asr.models.layers.attention import get_attention
+from espnet_onnx.export.asr.models.decoders.predecoder import PreDecoder
 from espnet_onnx.utils.abs_model import AbsExportModel
 from espnet_onnx.utils.torch_function import MakePadMask, normalize
