+++ espnet_onnx/export/tts/models/tts_models/vits.py
@@ -4,7 +4,7 @@ from typing import Optional, Tuple
 import torch
 import torch.nn as nn
 
-from espnet_onnx.export.asr.models.language_models.embed import get_pos_emb
+from espnet_onnx.export.asr.models.layers.embed import get_pos_emb
 from espnet_onnx.utils.abs_model import AbsExportModel
 from espnet_onnx.utils.torch_function import MakePadMask, normalize
